To be is a to be the value of a variable

+++ 0x10-variadic_functions/3-print_all.c
@@ -20,7 +20,7 @@ void print_all(const char * const format, ...)
 		switch (format[j])
 		{
 			case 'i':
-				printf("%d", (int) va_arg(ap, int));
+				printf("%d", va_arg(ap, int));
 				break;
 			case 'c':
 				printf("%c", (char) va_arg(ap, int));
